requirements.txt
torch>=2.0
tape-proteins
pandas
scikit-learn
biopython
numpy
